refactor(shadcn): rename EmbedPopupView_01 to AgentPopup_01
Renames the agents-ui popup block and everything tied to its old
name (directory, component, types, registry entry, tests, and
Storybook story) for a clearer, more consistent name.
