+++ packages/shadcn/components/agents-ui/blocks/agent-popup-01/components/popup-view.tsx
@@ -7,12 +7,12 @@ import { type AgentControlBarControls } from '@/components/agents-ui/agent-contr
 import { AgentSessionView_01 } from '@/components/agents-ui/blocks/agent-session-view-01/components/agent-session-block';
 import type { AudioVisualizerConfig } from '@/components/agents-ui/blocks/agent-session-view-01/components/agent-session-block';
 import { cn } from '@/lib/utils';
-import type { EmbedPopupViewError } from './embed-popup-block';
+import type { AgentPopupError } from './agent-popup-block';
 
 interface ErrorOverlayProps {
   logo?: string;
   agentName?: string;
-  error?: EmbedPopupViewError;
+  error?: AgentPopupError;
 }
 
 function ErrorOverlay({ logo, agentName, error }: ErrorOverlayProps) {
@@ -52,7 +52,7 @@ function ErrorOverlay({ logo, agentName, error }: ErrorOverlayProps) {
 export interface PopupViewProps {
   agentName?: string;
   logo?: string;
-  error?: EmbedPopupViewError;
+  error?: AgentPopupError;
   controls: AgentControlBarControls;
   themeMode?: 'dark' | 'light';
   preConnectMessage?: string;
