+++ packages/shadcn/components/agents-ui/blocks/agent-popup-01/index.ts
@@ -1,0 +1,2 @@
+export * from './components/agent-popup-block';
+export * from './components/trigger';
