--- packages/shadcn/components/agents-ui/blocks/embed-popup-view-01/index.ts
@@ -1,2 +1,0 @@
-export * from './components/embed-popup-block';
-export * from './components/trigger';
