+++ packages/shadcn/index.ts
@@ -13,4 +13,4 @@ export * from './components/agents-ui/agent-audio-visualizer-wave';
 export * from './components/agents-ui/agent-audio-visualizer-aura';
 export * from './components/agents-ui/react-shader-toy';
 export * from './components/agents-ui/blocks/agent-session-view-01/components/agent-session-block';
-export * from './components/agents-ui/blocks/embed-popup-view-01/components/embed-popup-block';
+export * from './components/agents-ui/blocks/agent-popup-01/components/agent-popup-block';
